renamed to com.examples.cloud.speech
