+++ speech/grpc/bin/speech-sample-streaming.sh
@@ -15,4 +15,4 @@
 
 SRC_DIR=$(cd "$(dirname "$0")/.."; pwd)
 java -cp ${SRC_DIR}/target/grpc-sample-1.0-jar-with-dependencies.jar \
-    com.google.cloud.speech.grpc.demos.StreamingRecognizeClient "$@"
+    com.examples.cloud.speech.StreamingRecognizeClient "$@"
